v8(services): Handle error case in action when getting share spaces
Also moved service instance sharing related client functions closer
to other service instance functions

[#171843321](https://www.pivotaltracker.com/story/show/171843321)
